No need for transition matrix
In this case you don't really need a transition matrix. In a realistic scenario you would not know it.
Isn't it better to pass the action straight to the step(action).
Basically commenting the line `action = np.random.choice(4, 1, p=self.transition_matrix[int(action),:])` in the `step()` method of gridworld?